pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docmcp"
version = "0.1.0"
description = "Enterprise Document Processing System with MCP Protocol Support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "DocMCP Team", email = "team@docmcp.dev"},
]
keywords = [
    "document-processing",
    "mcp",
    "model-context-protocol",
    "skills",
    "plugins",
    "async",
    "enterprise",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Enterprise",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business",
    "Typing :: Typed",
]

dependencies = [
    # Core async
    "aiofiles>=23.0.0",
    "aiohttp>=3.8.0",
    
    # Document processing
    "python-docx>=0.8.11",
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.9.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    
    # Data handling
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    
    # Security
    "cryptography>=41.0.0",
    "PyJWT>=2.8.0",
    
    # Utilities
    "watchdog>=3.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
    "pre-commit>=3.4.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
    "mkdocstrings[python]>=0.22.0",
]

server = [
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.0",
    "gunicorn>=21.2.0",
]

redis = [
    "redis>=5.0.0",
    "aioredis>=2.0.0",
]

database = [
    "asyncpg>=0.28.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.12.0",
]

all = [
    "docmcp[dev,docs,server,redis,database]",
]

[project.urls]
Homepage = "https://github.com/docmcp/docmcp"
Documentation = "https://docmcp.readthedocs.io"
Repository = "https://github.com/docmcp/docmcp.git"
"Bug Tracker" = "https://github.com/docmcp/docmcp/issues"

[project.scripts]
docmcp-server = "docmcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["docmcp"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["docmcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
